+++ src/ssw/Force/Common/CommonTools.java
@@ -31,16 +31,16 @@ import java.awt.Font;
 import java.io.File;
 
 public class CommonTools {
-    private final static float[][] BVMods = {
-        { 2.8f, 2.63f, 2.45f, 2.28f, 2.01f, 1.82f, 1.75f, 1.67f, 1.59f },
-        { 2.56f, 2.4f, 2.24f, 2.08f, 1.84f, 1.6f, 1.58f, 1.51f, 1.44f },
-        { 2.24f, 2.1f, 1.96f, 1.82f, 1.61f, 1.4f, 1.33f, 1.31f, 1.25f },
-        { 1.92f, 1.8f, 1.68f, 1.56f, 1.38f, 1.2f, 1.14f, 1.08f, 1.06f },
-        { 1.6f, 1.5f, 1.4f, 1.3f, 1.15f, 1.0f, 0.95f, 0.9f, 0.85f },
-        { 1.5f, 1.35f, 1.26f, 1.17f, 1.04f, 0.9f, 0.86f, 0.81f, 0.77f },
-        { 1.43f, 1.33f, 1.19f, 1.11f, 0.98f, 0.85f, 0.81f, 0.77f, 0.72f },
-        { 1.36f, 1.26f, 1.16f, 1.04f, 0.92f, 0.8f, 0.76f, 0.72f, 0.68f },
-        { 1.28f, 1.19f, 1.1f, 1.01f, 0.86f, 0.75f, 0.71f, 0.68f, 0.64f }
+    private final static double[][] BVMods = {
+        { 2.8, 2.63, 2.45, 2.28, 2.01, 1.82, 1.75, 1.67, 1.59 },
+        { 2.56, 2.4, 2.24, 2.08, 1.84, 1.6, 1.58, 1.51, 1.44 },
+        { 2.24, 2.1, 1.96, 1.82, 1.61, 1.4, 1.33, 1.31, 1.25 },
+        { 1.92, 1.8, 1.68, 1.56, 1.38, 1.2, 1.14, 1.08, 1.06 },
+        { 1.6, 1.5, 1.4, 1.3, 1.15, 1.0, 0.95, 0.9, 0.85 },
+        { 1.5, 1.35, 1.26, 1.17, 1.04, 0.9, 0.86, 0.81, 0.77 },
+        { 1.43, 1.33, 1.19, 1.11, 0.98, 0.85, 0.81, 0.77, 0.72 },
+        { 1.36, 1.26, 1.16, 1.04, 0.92, 0.8, 0.76, 0.72, 0.68 },
+        { 1.28, 1.19, 1.1, 1.01, 0.86, 0.75, 0.71, 0.68, 0.64 }
     };
 
     public final static String tab = "    ";
@@ -52,25 +52,25 @@ public class CommonTools {
     public static Font TitleFont = new Font("Arial", Font.BOLD, 14);
     public static Font SectionHeaderFont = new Font("Arial", Font.BOLD, 12);
 
-    public static float GetSkillBV( float BV, int Gunnery, int Piloting ) {
+    public static double GetSkillBV( double BV, int Gunnery, int Piloting ) {
         return BV * BVMods[Gunnery][Piloting];
     }
 
-    public static float GetModifierBV( float SkillBV, float Modifier) {
+    public static double GetModifierBV( double SkillBV, double Modifier) {
         return SkillBV * Modifier;
     }
 
-    public static float GetFullAdjustedBV( float BV, int Gunnery, int Piloting, float Modifier ) {
+    public static double GetFullAdjustedBV( double BV, int Gunnery, int Piloting, double Modifier ) {
         return BV * BVMods[Gunnery][Piloting] * Modifier;
     }
 
-    public static float GetForceSizeMultiplier( int Force1Size, int Force2Size )
+    public static double GetForceSizeMultiplier( int Force1Size, int Force2Size )
     {
         if( Force1Size <= 0 || Force2Size <= 0 ) {
             return 0;
         }
 
-        return ( ((float) Force2Size) / ((float) Force1Size) ) + ( ((float) Force1Size) / ((float) Force2Size) ) - 1.0f;
+        return ( ((double) Force2Size) / ((double) Force1Size) ) + ( ((double) Force1Size) / ((double) Force2Size) ) - 1.0f;
     }
 
     public static String SafeFileName(String filename) {
